Revert "Merge branch 'master' into textarea-autoresize"
This reverts commit 25fddd53e4998c85af5ac4afc01e3c45f7989017, reversing
changes made to e39b767528c9fcdc5cd0e71f70164fd9d8cc199d.

+++ application/classes/Controller/Comments.php
@@ -4,84 +4,56 @@ class Controller_Comments extends Controller_Base_preDispatch
 {
     public function action_add()
     {
+        $comment = new Model_Comment();
+
+        $page = new Model_Page($this->request->param('id'));
+
+        $text = trim(nl2br(Arr::get($_POST, 'add_comment_textarea')));
+
         $error = '';
 
         /**
-         * Checking for CSRF
+         * Checking for existing page
          */
-        $csrfToken = Arr::get($_POST, 'csrf');
-        if (!Security::check($csrfToken)) {
-            $error = 'Ошибка валидации CSRF-токена';
+        if (!$page->id) {
+
+            $error = 'Wrong page id';
             goto finish;
         }
-        /***/
 
         /**
          * Checking for authorized user
          */
         if ($this->user->status < Model_User::USER_STATUS_REGISTERED) {
-            $error = 'Ошибка доступа';
-            goto finish;
-        }
-        /***/
 
-        /**
-         * Checking for existing page
-         */
-        $page = new Model_Page($this->request->param('id'));
-        if (!$page->id) {
-            $error = 'Неверный номер страницы';
+            $error = 'Access denied';
             goto finish;
         }
-        /***/
 
         /**
          * Checking for existing text
          */
-        $text = trim(nl2br(Arr::get($_POST, 'comment_text')));
         if (!$text) {
-            $error = 'Отсутствует текст комментария';
+
+            $error = 'Text is incorrect';
             goto finish;
         }
-        /***/
 
-        $comment = new Model_Comment();
-        $comment->page_id              = $page->id;
+        $comment->page_id              = $this->request->param('id');
         $comment->text                 = $text;
         $comment->parent_comment['id'] = Arr::get($_POST, 'parent_id', '0');
         $comment->author['id']         = $this->user->id;
         $comment->root_id              = Arr::get($_POST, 'root_id', '0');
-        $comment_id = $comment->insert();
 
-        $comment = Model_Comment::get($comment_id);
-        $comment->parent_comment = $comment->parent_id ? Model_Comment::get($comment->parent_id) : null;
+        $comment->insert();
 
         finish:
-
-        if (Model_Methods::isAjax()) {
-
-            $response = array();
-
-            if ($error) {
-                $response['success'] = 0;
-                $response['error'] = $error;
-            } else {
-                $response['success'] = 1;
-                $response['comment'] = View::factory('templates/comments/comment', array('user' => $this->user, 'comment' => $comment, 'isOnPage' => true))->render();
-                $response['commentId'] = $comment->id;
-            }
-
-            $this->auto_render = false;
-            $this->response->headers('Content-Type', 'application/json; charset=utf-8');
-            $this->response->body( json_encode($response) );
-
-            return true;
-
-        }
-
         if ($error) {
+
             $this->redirect( '/p/' . $page->id . '/' . $page->uri . '?error=' . $error );
+
         } else {
+
             $this->redirect( '/p/' . $page->id . '/' . $page->uri );
         }
     }
@@ -92,9 +64,7 @@ class Controller_Comments extends Controller_Base_preDispatch
 
         $comment = Model_Comment::get($comment_id);
 
-        if ($comment->author->id == $this->user->id || $this->user->isAdmin) {
-            $comment->delete(true);
-        }
+        if ($comment->author->id == $this->user->id || $this->user->isAdmin) $comment->delete(true);
 
         $page = new Model_Page($comment->page_id);
